requirements.txt
# Core libraries for Agent, MCP, and Telegram
openai-agents
fastmcp
python-telegram-bot

# Utility libraries
python-dotenv
requests
openai
